Pete's Commit
I added a (bad) background, a black platform, and set the player.png to
a rigidbody with a 2d box collider that fits around the white square
itself. Terrain consists of a black box, the camera is in an
orthographic position as well. Also added PlayerTerrain layer where
player and terrain interact.