added Death and Animations
Now should restart loaded level upon death, we can change that to instead load a death level, and also added some animations. I don't know if they will be download right. I added some code to the movement so it flips the entire character so the animations would be easier and also triggers the animations.